moved dayBetween and weekdayBetween to dateCalculation
